add the ability to mount/unmount the whole app
