+++ index.html
@@ -36,6 +36,6 @@
       <div class="demo-1-box demo-1-selected">Selected</div>
     </div>
 
-    <script src="src/index.js"></script>
+    <script src="./src/index.js"></script>
   </body>
 </html>
